% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ITconverter.R
\name{ITconverter}
\alias{ITconverter}
\title{Converts IT measures to body mass, foraging distance and tongue length for bees.}
\usage{
ITconverter(IT, family, type = "all", mouthpart = "all")
}
\arguments{
\item{IT}{A vector of bee intertegular spans (IT) measurments in cm.}

\item{family}{a vector of bee families. Only implemented 5 out of the extant 7 families:
"Andrenidae", "Apidae", "Colletidae", "Halictidae", "Megachilidae".}

\item{type}{The type of foraging distance desired. Options are "all", "Maximum homing distance",
"Typical homing distance", "Maximum feeder training distance", "Maximum communication distance". See details in Greenleaf et al. 2007.}

\item{mouthpart}{The mouthpart you are interested in. Options are "all", glossa", "prementum" and "tongue" (i.e. gloss + prementum)}
}
\value{
A dataframe with bee body masses (gr), tongue length (mm) and foraging distance
(km) is returned for each bees species.
}
\description{
Calculates body mass form cane (1987), foraging distande from Greenleaf et al. (2007) and
tongue length from Cariveau et al. (2015) using intertegular distance values (IT).
}

