% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ITweigth.R
\name{ITweigth}
\alias{ITweigth}
\title{Converts IT measures to body mass for bees.}
\usage{
ITweigth(IT)
}
\arguments{
\item{IT}{A vector of bee intertegular spans (IT) measurments in cm.}
}
\value{
A dataframe with bee body masses (gr) is returned for each bees species.
}
\description{
Calculates body mass form cane (1987), using intertegular distance values (IT).
}

