% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ITconverter.R
\name{weigth_fn}
\alias{weigth_fn}
\title{Converts IT measures to body mass, foraging distance and tongue length for bees.}
\usage{
weigth_fn(IT)
}
\arguments{
\item{IT}{A vector of bee intertegular spans (IT) measurments in cm.}

\item{family}{a vector of bee families. Only implemented 5 out of the extant 7 families:
"Andrenidae", "Apidae", "Colletidae", "Halictidae", "Megachilidae".}
}
\value{
A dataframe with bee body masses (gr), tongue length (mm) and foraging distance
(m) is returned for each bees species.
}
\description{
Calculates body mass form cane (1987), foraging distande from Greenleaf et al. (2007) and
tongue length from Cariveau et al. (2015) using intertegular distance values (IT).
}
\examples{
x <- rnorm(100, 10, 2)
f <- rep(c("Andrenidae", "Apidae", "Colletidae", "Halictidae", "Megachilidae"),20)
y <- ITconverter(IT = x, family = f)
plot(log(x)~log(y$body_mass))
plot(log(x)~log(y$tongue_length))
plot(y$tongue_length ~ y$body_mass)
}

