% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/tongues.R
\docType{data}
\name{tongues}
\alias{tongues}
\title{Bee tongue length measurments}
\format{A data frame with 101 rows and 11 variables}
\usage{
data(tongues)
}
\description{
A dataset containing the mean bee tongue length for >100 species. The variables are as follows:
}
\details{
\itemize{
  \item Family.
  \item genus.
  \item species.
  \item tongue_type.
  \item count.
  \item tongue.m.
  \item ITlength_mm.m.
  \item glossa_mm.m.
  \item prementum_mm.m.
  \item head_width.m.
  \item resid_tongue.
  \item resid_glossa.
  }
}
\keyword{datasets}

